+++ jwst/datamodels/ramp.py
@@ -1,3 +1,5 @@
+import warnings
+
 from .model_base import DataModel
 
 
@@ -41,3 +43,10 @@ class RampModel(DataModel):
         self.pixeldq = self.pixeldq
         self.groupdq = self.groupdq
         self.err = self.err
+
+
+def MIRIRampModel(*args, **kwargs):
+    warnings.simplefilter('default')
+    warnings.warn(message="MIRIRampModel is deprecated and will be removed.  "
+        "Use RampModel.", category=DeprecationWarning)
+    return RampModel(*args, **kwargs)