+++ jwst/jump/tests/test_detect_jumps.py
@@ -2,7 +2,7 @@ import numpy as np
 import pytest
 
 from jwst.datamodels import GainModel, ReadnoiseModel
-from jwst.datamodels import MIRIRampModel
+from jwst.datamodels import RampModel
 from jwst.jump.jump import detect_jumps
 
 
@@ -102,7 +102,7 @@ def setup_inputs():
         pixdq = np.zeros(shape=(nrows, ncols), dtype=np.float64)
         read_noise = np.full((nrows, ncols), readnoise, dtype=np.float64)
         gdq = np.zeros(shape=(nints, ngroups, nrows, ncols), dtype=np.int32)
-        model1 = MIRIRampModel(data=data, err=err, pixeldq=pixdq, groupdq=gdq, times=times)
+        model1 = RampModel(data=data, err=err, pixeldq=pixdq, groupdq=gdq, times=times)
         model1.meta.instrument.name = 'MIRI'
         model1.meta.instrument.detector = 'MIRIMAGE'
         model1.meta.instrument.filter = 'F480M'